+++ verik-compiler/src/main/kotlin/io/verik/compiler/common/Visitor.kt
@@ -48,6 +48,7 @@ import io.verik.compiler.ast.element.declaration.sv.EAlwaysComBlock
 import io.verik.compiler.ast.element.declaration.sv.EAlwaysSeqBlock
 import io.verik.compiler.ast.element.declaration.sv.EClockingBlock
 import io.verik.compiler.ast.element.declaration.sv.EClockingBlockInstantiation
+import io.verik.compiler.ast.element.declaration.sv.ECluster
 import io.verik.compiler.ast.element.declaration.sv.EComponentInstantiation
 import io.verik.compiler.ast.element.declaration.sv.EConstraint
 import io.verik.compiler.ast.element.declaration.sv.ECoverBin
@@ -55,7 +56,6 @@ import io.verik.compiler.ast.element.declaration.sv.ECoverCross
 import io.verik.compiler.ast.element.declaration.sv.ECoverGroup
 import io.verik.compiler.ast.element.declaration.sv.ECoverPoint
 import io.verik.compiler.ast.element.declaration.sv.EEnum
-import io.verik.compiler.ast.element.declaration.sv.EGenerateForBlock
 import io.verik.compiler.ast.element.declaration.sv.EInitialBlock
 import io.verik.compiler.ast.element.declaration.sv.EInjectedProperty
 import io.verik.compiler.ast.element.declaration.sv.EModule
@@ -338,8 +338,8 @@ abstract class Visitor {
         visitAbstractProperty(constraint)
     }
 
-    open fun visitGenerateForBlock(generateForBlock: EGenerateForBlock) {
-        visitAbstractProperty(generateForBlock)
+    open fun visitCluster(cluster: ECluster) {
+        visitAbstractProperty(cluster)
     }
 
     open fun visitAbstractValueParameter(abstractValueParameter: EAbstractValueParameter) {