+++ verik-compiler/src/main/kotlin/io/verik/compiler/interpret/FileSplitterStage.kt
@@ -21,9 +21,9 @@ import io.verik.compiler.ast.element.declaration.common.EEnumEntry
 import io.verik.compiler.ast.element.declaration.common.EFile
 import io.verik.compiler.ast.element.declaration.common.EProperty
 import io.verik.compiler.ast.element.declaration.sv.EAbstractComponent
+import io.verik.compiler.ast.element.declaration.sv.ECluster
 import io.verik.compiler.ast.element.declaration.sv.ECoverGroup
 import io.verik.compiler.ast.element.declaration.sv.EEnum
-import io.verik.compiler.ast.element.declaration.sv.EGenerateForBlock
 import io.verik.compiler.ast.element.declaration.sv.EInjectedProperty
 import io.verik.compiler.ast.element.declaration.sv.EStruct
 import io.verik.compiler.ast.element.declaration.sv.ESvClass
@@ -95,7 +95,7 @@ object FileSplitterStage : ProjectStage() {
             is EInjectedProperty -> false
             is EProperty -> false
             is EEnumEntry -> false
-            is EGenerateForBlock -> isRootPackageDeclaration(declaration.declaration)
+            is ECluster -> isRootPackageDeclaration(declaration.declarations[0])
             else -> Messages.INTERNAL_ERROR.on(
                 declaration,
                 "Unexpected declaration type: ${declaration::class.simpleName}"