+++ verik-compiler/src/main/kotlin/io/verik/compiler/serialize/source/DeclarationSerializer.kt
@@ -30,7 +30,6 @@ import io.verik.compiler.ast.element.declaration.sv.ECoverCross
 import io.verik.compiler.ast.element.declaration.sv.ECoverGroup
 import io.verik.compiler.ast.element.declaration.sv.ECoverPoint
 import io.verik.compiler.ast.element.declaration.sv.EEnum
-import io.verik.compiler.ast.element.declaration.sv.EGenerateForBlock
 import io.verik.compiler.ast.element.declaration.sv.EInitialBlock
 import io.verik.compiler.ast.element.declaration.sv.EInjectedProperty
 import io.verik.compiler.ast.element.declaration.sv.EModule
@@ -366,29 +365,6 @@ object DeclarationSerializer {
         }
         serializeContext.label(constraint.body.endLocation) {
             serializeContext.appendLine("}")
-        }
-    }
-
-    fun serializeGenerateForBlock(generateForBlock: EGenerateForBlock, serializeContext: SerializeContext) {
-        fun serializeBody() {
-            val indexName = generateForBlock.indexProperty.name
-            serializeContext.append("for (genvar $indexName = 0; $indexName < ${generateForBlock.size}; $indexName++)")
-            serializeContext.appendLine(" begin : ${generateForBlock.name}")
-            serializeContext.indent {
-                serializeContext.serialize(generateForBlock.declaration)
-            }
-            serializeContext.label(generateForBlock.endLocation) {
-                serializeContext.appendLine("end : ${generateForBlock.name}")
-            }
-        }
-        if (generateForBlock.parent is EGenerateForBlock) {
-            serializeBody()
-        } else {
-            serializeContext.appendLine("generate")
-            serializeContext.indent { serializeBody() }
-            serializeContext.label(generateForBlock.endLocation) {
-                serializeContext.appendLine("endgenerate")
-            }
         }
     }